Moved delegate-selector and delegate-connectivity under ui package folder. Reuse them in apps/subjects/views.
