+++ apps/design-system/src/subjects/views/delegates/delegate-selector.tsx
@@ -1,146 +1,21 @@
-import { useState } from 'react'
+import { DelegateSelector as DelegateSelectorComponent } from '@harnessio/ui/views';
 
-import { useTranslationStore } from '@utils/viewUtils'
-import { defaultTo } from 'lodash-es'
+import mockTagsList from './mock-tags-list.json';
 
-import { Drawer, FormSeparator, Icon, StyledLink } from '@harnessio/ui/components'
-import {
-  DelegateSelectionTypes,
-  DelegateSelectorForm,
-  DelegateSelectorFormFields,
-  DelegateSelectorInput
-} from '@harnessio/ui/views'
+import { getMatchedDelegatesCount, isDelegateSelected } from '@subjects/views/delegates/utils.ts'
+import { useTranslationStore } from '@utils/viewUtils.ts'
+import { useDelegateData } from '@subjects/views/delegates/hooks'
 
-import mockDelegatesList from './mock-delegates-list.json'
-import { getMatchedDelegatesCount, isDelegateSelected } from './utils'
-
-const delegatesData = mockDelegatesList.map(delegate => ({
-  groupId: delegate.groupId,
-  groupName: delegate.groupName,
-  lastHeartBeat: delegate.lastHeartBeat,
-  activelyConnected: delegate.activelyConnected,
-  groupCustomSelectors: delegate.groupCustomSelectors || [],
-  groupImplicitSelectors: [...Object.keys(defaultTo(delegate.groupImplicitSelectors, {}))]
-}))
-
-const mockTagsList = [
-  'sanity-windows',
-  'eightfivetwoold',
-  'qa-automation',
-  'sanity',
-  'self-hosted-vpc-delegate',
-  'local',
-  '_testDocker',
-  'myrunner',
-  'macos-arm64',
-  'west1-delegate-qa',
-  'linux-amd64',
-  'eightfivetwo',
-  'automation-eks-delegate'
-]
-
-const renderSelectedValue = (type: DelegateSelectionTypes | null, tags: string[]) =>
-  type === DelegateSelectionTypes.TAGS ? tags.join(', ') : type === DelegateSelectionTypes.ANY ? 'any delegate' : null
-
-/* ----------  DRAWER COMPONENT  -------------- */
-interface DrawerProps {
-  open: boolean
-  setOpen: (open: boolean) => void
-  preSelectedTags: string[]
-  onSubmit: (data: DelegateSelectorFormFields) => void
-  disableAnyDelegate?: boolean
-}
-
-const DelegateSelectorDrawer = ({ open, setOpen, preSelectedTags, onSubmit, disableAnyDelegate }: DrawerProps) => (
-  <Drawer.Root open={open} onOpenChange={setOpen} direction="right">
-    <Drawer.Content className="w-1/2">
-      <Drawer.Header>
-        <Drawer.Title className="text-cn-foreground-1 mb-2 text-xl">Delegate selector</Drawer.Title>
-        <FormSeparator className="w-full" />
-        <div className="flex">
-          Haven&apos;t installed a delegate yet?
-          <StyledLink className="flex flex-row items-center ml-1" variant="accent" to="#">
-            Install delegate <Icon name="attachment-link" className="ml-1" size={12} />
-          </StyledLink>
-        </div>
-        <Drawer.Close onClick={() => setOpen(false)} />
-      </Drawer.Header>
-
-      <DelegateSelectorForm
-        delegates={delegatesData}
-        tagsList={mockTagsList}
-        useTranslationStore={useTranslationStore}
-        isLoading={false}
-        onFormSubmit={onSubmit}
-        onBack={() => setOpen(false)}
-        isDelegateSelected={isDelegateSelected}
-        getMatchedDelegatesCount={getMatchedDelegatesCount}
-        preSelectedTags={preSelectedTags}
-        disableAnyDelegate={disableAnyDelegate}
-      />
-    </Drawer.Content>
-  </Drawer.Root>
-)
-
-/* ----------  MAIN COMPONENT  -------------------------- */
 export const DelegateSelector = () => {
-  /* ---- FIRST (ANY allowed) ---- */
-  const [openA, setOpenA] = useState(false)
-  const [typeA, setTypeA] = useState<DelegateSelectionTypes | null>(null)
-  const [tagsA, setTagsA] = useState<string[]>([])
-
-  /* ---- SECOND (ANY disabled) --- */
-  const [openB, setOpenB] = useState(false)
-  const [typeB, setTypeB] = useState<DelegateSelectionTypes | null>(null)
-  const [tagsB, setTagsB] = useState<string[]>([])
-
-  const handleSubmitA = ({ type, tags }: DelegateSelectorFormFields) => {
-    setTypeA(type === DelegateSelectionTypes.ANY ? DelegateSelectionTypes.ANY : DelegateSelectionTypes.TAGS)
-    setTagsA(type === DelegateSelectionTypes.TAGS ? tags.map(t => t.id) : [])
-    setOpenA(false)
-  }
-
-  const handleSubmitB = ({ tags }: DelegateSelectorFormFields) => {
-    setTypeB(DelegateSelectionTypes.TAGS)
-    setTagsB(tags.map(t => t.id))
-    setOpenB(false)
-  }
+  const delegatesData = useDelegateData();
 
   return (
-    <div className="p-5">
-      <DelegateSelectorInput
-        placeholder={<StyledLink to="#"> select a delegate</StyledLink>}
-        value={renderSelectedValue(typeA, tagsA)}
-        label="Delegate selector"
-        onClick={() => setOpenA(true)}
-        onEdit={() => setOpenA(true)}
-        onClear={() => setTagsA([])}
-        renderValue={tag => tag}
-        className="max-w-xs mb-8"
-      />
-
-      <DelegateSelectorDrawer open={openA} setOpen={setOpenA} preSelectedTags={tagsA} onSubmit={handleSubmitA} />
-
-      <div className="pt-10">
-        <DelegateSelectorInput
-          placeholder={<StyledLink to="#">select a delegate (any disabled)</StyledLink>}
-          value={renderSelectedValue(typeB, tagsB)}
-          label="Delegate selector"
-          onClick={() => setOpenB(true)}
-          onEdit={() => setOpenB(true)}
-          onClear={() => setTagsB([])}
-          renderValue={tag => tag}
-          className="max-w-xs mb-8"
-        />
-
-        <DelegateSelectorDrawer
-          open={openB}
-          setOpen={setOpenB}
-          preSelectedTags={tagsB}
-          onSubmit={handleSubmitB}
-          disableAnyDelegate
-        />
-      </div>
-    </div>
+    <DelegateSelectorComponent
+      delegates={delegatesData}
+      getMatchedDelegatesCount={getMatchedDelegatesCount}
+      isDelegateSelected={isDelegateSelected}
+      tagsList={mockTagsList}
+      useTranslationStore={useTranslationStore}
+    />
   )
 }
