+++ apps/design-system/src/subjects/views/delegates/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-delegates.ts';
