+++ packages/ui/src/views/delegates/delegate-connectivity/index.ts
@@ -1,0 +1,3 @@
+export {
+  DelegateConnectivityWrapper
+} from './delegate-connectivity-wrapper';
