+++ packages/ui/src/views/delegates/delegate-selector/index.ts
@@ -1,0 +1,1 @@
+export { DelegateSelector } from './delegate-selector';
