+++ packages/ui/src/views/delegates/index.ts
@@ -1,4 +1,3 @@
 export * from './types'
-export * from './components/delegate-connectivity-list'
-export * from './delegate-selector/delegate-selector-form'
-export * from './delegate-selector/delegate-selector-input'
+export * from './delegate-selector'
+export * from './delegate-connectivity'
